requirements.txt
# python >= 3.10
numpy
opencv-python
gdal
pyyaml
tqdm